large_file_upload azure: update code so that filename can be changed (#37)
* large_file_upload azure: update code so that filename can be changed

Updated the code so that when the cloud service provider is set to "azure",
the filename is updated via the env.default request instead of the filename
parameter in web.put/web.post.

* large_file_upload azure: optimize azure filename env variable sync

Updated the code to call `hub.sync` only when necessary. The filename env
variable is now synced only if the Notehub's filename env variable differs
from the local one, avoiding unnecessary sync calls.This change was made
based on the PR feedback.